backend/requirements.txt
google-generativeai
fastapi
uvicorn
opencv-python-headless
numpy
torch
torchvision
Pillow
python-multipart
scikit-image
python-dotenv
